Move index handlers into handler package
