+++ src/main/java/com/yelp/nrtsearch/server/grpc/LuceneServer.java
@@ -51,6 +51,7 @@ import com.yelp.nrtsearch.server.luceneserver.handler.GetStateHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.GlobalStateHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.IndicesHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.LiveSettingsHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.LiveSettingsV2Handler;
 import com.yelp.nrtsearch.server.luceneserver.handler.MetricsHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.NewNRTPointHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.ReadyHandler;
@@ -66,6 +67,7 @@ import com.yelp.nrtsearch.server.luceneserver.handler.SearchHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.SearchV2Handler;
 import com.yelp.nrtsearch.server.luceneserver.handler.SendRawFileHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.SettingsHandler;
+import com.yelp.nrtsearch.server.luceneserver.handler.SettingsV2Handler;
 import com.yelp.nrtsearch.server.luceneserver.handler.StartIndexHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.StartIndexV2Handler;
 import com.yelp.nrtsearch.server.luceneserver.handler.StatsHandler;
@@ -74,8 +76,6 @@ import com.yelp.nrtsearch.server.luceneserver.handler.StopIndexHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.UpdateFieldsHandler;
 import com.yelp.nrtsearch.server.luceneserver.handler.WriteNRTPointHandler;
 import com.yelp.nrtsearch.server.luceneserver.highlights.HighlighterService;
-import com.yelp.nrtsearch.server.luceneserver.index.handlers.LiveSettingsV2Handler;
-import com.yelp.nrtsearch.server.luceneserver.index.handlers.SettingsV2Handler;
 import com.yelp.nrtsearch.server.luceneserver.logging.HitsLoggerCreator;
 import com.yelp.nrtsearch.server.luceneserver.rescore.RescorerCreator;
 import com.yelp.nrtsearch.server.luceneserver.script.ScriptService;
