+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/handler/RegisterFieldsHandler.java
@@ -18,7 +18,6 @@ package com.yelp.nrtsearch.server.luceneserver.handler;
 import com.yelp.nrtsearch.server.grpc.FieldDefRequest;
 import com.yelp.nrtsearch.server.grpc.FieldDefResponse;
 import com.yelp.nrtsearch.server.luceneserver.index.IndexStateManager;
-import com.yelp.nrtsearch.server.luceneserver.index.handlers.FieldUpdateHandler;
 import com.yelp.nrtsearch.server.luceneserver.state.GlobalState;
 import io.grpc.Status;
 import io.grpc.stub.StreamObserver;
@@ -40,7 +39,8 @@ public class RegisterFieldsHandler extends Handler<FieldDefRequest, FieldDefResp
     try {
       IndexStateManager indexStateManager =
           getGlobalState().getIndexStateManager(fieldDefRequest.getIndexName());
-      FieldDefResponse reply = FieldUpdateHandler.handle(indexStateManager, fieldDefRequest);
+      String updatedFields = indexStateManager.updateFields(fieldDefRequest.getFieldList());
+      FieldDefResponse reply = FieldDefResponse.newBuilder().setResponse(updatedFields).build();
       logger.info("RegisterFieldsHandler registered fields " + reply);
       responseObserver.onNext(reply);
       responseObserver.onCompleted();
